+++ prompt.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 import argparse
 import datetime
@@ -8,6 +8,7 @@ import psutil
 import traceback
 import logging
 
+from logzero import logger
 from prompt_toolkit import prompt
 from prompt_toolkit.contrib.completers import WordCompleter
 from prompt_toolkit.history import FileHistory
@@ -44,6 +45,9 @@ from neo.UserPreferences import preferences
 from neocore.KeyPair import KeyPair
 from neocore.UInt256 import UInt256
 
+from neorpc.Client import RPCClient
+from neorpc.Settings import settings as rpc_settings
+
 # Logfile settings & setup
 LOGFILE_FN = os.path.join(DIR_PROJECT_ROOT, 'prompt.log')
 LOGFILE_MAX_BYTES = 5e7  # 50 MB
@@ -52,6 +56,10 @@ settings.set_logfile(LOGFILE_FN, LOGFILE_MAX_BYTES, LOGFILE_BACKUP_COUNT)
 
 # Prompt history filename
 FILENAME_PROMPT_HISTORY = os.path.join(DIR_PROJECT_ROOT, '.prompt.py.history')
+
+
+class PrivnetConnectionError(Exception):
+    pass
 
 
 class PromptInterface(object):
@@ -84,6 +92,7 @@ class PromptInterface(object):
                 'import nep2 {nep2_encrypted_key}',
                 'import contract {path/to/file.avm} {params} {returntype} {needs_storage} {needs_dynamic_invoke}',
                 'import contract_addr {contract_hash} {pubkey}',
+                'import multisig_addr {pubkey in wallet} {minimum # of signatures required} {signing pubkey 1} {signing pubkey 2}...',
                 'import watch_addr {address}',
                 'import token {token_contract_hash}',
                 'export wif {address}',
@@ -818,8 +827,21 @@ class PromptInterface(object):
             else:
                 print("Cannot configure log. Please specify on|off")
 
+        elif what == 'sc-debug-notify':
+            c1 = get_arg(args, 1).lower()
+            if c1 is not None:
+                if c1 == 'on' or c1 == '1':
+                    print("Smart contract emit Notify events on execution failure is now enabled")
+                    settings.set_emit_notify_events_on_sc_execution_error(True)
+                if c1 == 'off' or c1 == '0':
+                    print("Smart contract emit Notify events on execution failure is now disabled")
+                    settings.set_emit_notify_events_on_sc_execution_error(False)
+
+            else:
+                print("Cannot configure log. Please specify on|off")
+
         else:
-            print("Cannot configure %s try 'config sc-events on|off' or 'config debug on|off'", what)
+            print("Cannot configure %s try 'config sc-events on|off', 'config debug on|off' or 'sc-debug-notify on|off'" % what)
 
     def run(self):
         dbloop = task.LoopingCall(Blockchain.Default().PersistBlocks)
@@ -920,9 +942,35 @@ class PromptInterface(object):
                 traceback.print_exc()
 
 
+def check_privatenet():
+    """ Check if privatenet is running, and if container is same as the chain file """
+    rpc_settings.setup(settings.RPC_LIST)
+    client = RPCClient()
+    version = client.get_version()
+    if not version:
+        raise PrivnetConnectionError("Error: private network container doesn't seem to be running, or RPC is not enabled.")
+
+    print("Privatenet useragent '%s', nonce: %s" % (version["useragent"], version["nonce"]))
+
+    # Now check if nonce is the same as in the chain path
+    nonce_container = str(version["nonce"])
+    neopy_chain_meta_filename = os.path.join(settings.LEVELDB_PATH, ".privnet-nonce")
+    if os.path.isfile(neopy_chain_meta_filename):
+        nonce_chain = open(neopy_chain_meta_filename, "r").read()
+        if nonce_chain != nonce_container:
+            raise PrivnetConnectionError(
+                "Chain database in Chains/privnet is for a different private network than the current container. "
+                "Consider deleting the Chain directory with 'rm -rf Chains/privnet*'."
+            )
+    else:
+        with open(neopy_chain_meta_filename, "w") as f:
+            f.write(nonce_container)
+
+
 def main():
     parser = argparse.ArgumentParser()
 
+    # Network group
     group = parser.add_mutually_exclusive_group()
     group.add_argument("-m", "--mainnet", action="store_true", default=False,
                        help="Use MainNet instead of the default TestNet")
@@ -932,9 +980,16 @@ def main():
                        help="Use the CoZ network instead of the default TestNet")
     group.add_argument("-c", "--config", action="store", help="Use a specific config file")
 
+    # Theme
     parser.add_argument("-t", "--set-default-theme", dest="theme",
                         choices=["dark", "light"],
                         help="Set the default theme to be loaded from the config file. Default: 'dark'")
+
+    # Verbose
+    parser.add_argument("-v", "--verbose", action="store_true", default=False,
+                        help="Show smart-contract events by default")
+
+    # Show the neo-python version
     parser.add_argument("--version", action="version",
                         version="neo-python v{version}".format(version=__version__))
 
@@ -952,6 +1007,17 @@ def main():
 
     if args.theme:
         preferences.set_theme(args.theme)
+
+    if args.verbose:
+        settings.set_log_smart_contract_events(True)
+
+    # If privnet, check if correct chain
+    if args.privnet:
+        try:
+            check_privatenet()
+        except PrivnetConnectionError as e:
+            logger.error(str(e))
+            return
 
     # Instantiate the blockchain and subscribe to notifications
     blockchain = LevelDBBlockchain(settings.LEVELDB_PATH)