+++ CHANGELOG.rst
@@ -13,6 +13,9 @@ All notable changes to this project are documented in this file.
 - Renamed ``neo.api.REST.NotificationRestApi`` to ``neo.api.REST.RestApi``
 - Added ``-v/--verbose`` argument to prompt.py, which makes prompt.py show smart contract events by default
 - Added ``vm-log`` option to the ``config`` console command. This enabled logging of VM instructions to ``vm_instructions.log`` for debugging purposes.
+- prompt.py: When using a privnet with ``-p``, check if chain database is correct. Renamed ``Chains/Priv_Notif`` to ``Chains/privnet_notif`` (if you need your old privnet notification db, you need to rename it manually).
+- Fixed issue with missing ``addr`` location in ``neo/api/REST/RestApi.py``
+
 
 [0.5.3] 2018-03-04
 ------------------