+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ All notable changes to this project are documented in this file.
 - Fix multi-signature contract import to allow using a single signature
 - Fix fund sending from multi-signature contract
 - Added instructions on retrieving NEO TestNet funds
+- prompt.py: When using a privnet with ``-p``, check if chain database is correct. Renamed ``Chains/Priv_Notif`` to ``Chains/privnet_notif`` (if you need your old privnet notification db, you need to rename it manually).
+- Fixed issue with missing ``addr`` location in ``neo/api/REST/RestApi.py``
 
 
 [0.5.3] 2018-03-04