+++ CHANGELOG.rst
@@ -3,13 +3,31 @@ Changelog
 
 All notable changes to this project are documented in this file.
 
-[0.5.2] 2018-03-04
+[0.5.4-dev] Work in progress
+----------------------------
+- All requests to the API that are invalid will now receive a ``None`` for results rather than an empty list ``[]``
+- update to neo-boa==0.3.7
+- `api-server.py <https://github.com/CityOfZion/neo-python/blob/development/api-server.py>`_: Improved logging setup. See the options with ``./api-server.py -h``
+- Updated Dockerfile with command line parameters and extra --bootstrap option. Please check ./docker/README.md on how to use.
+- Added ``sc-debug-notify`` option to the ``config`` console command. This preserves smart contract ``Notify()`` events when SC execution fails and is intended for SC debugging purposes only.
+- Added VM instruction counter to ``ExecutionEngine.py`` error messages to indicate the final instruction that failed. Allows for setting conditional breakpoints to support SC debugging.
+- Renamed ``neo.api.REST.NotificationRestApi`` to ``neo.api.REST.RestApi``
+- Added ``-v/--verbose`` argument to prompt.py, which makes prompt.py show smart contract events by default
+- Fix multi-signature contract import to allow using a single signature
+- Fix fund sending from multi-signature contract
+- Added instructions on retrieving NEO TestNet funds
+- prompt.py: When using a privnet with ``-p``, check if chain database is correct. Renamed ``Chains/Priv_Notif`` to ``Chains/privnet_notif`` (if you need your old privnet notification db, you need to rename it manually).
+- Fixed issue with missing ``addr`` location in ``neo/api/REST/RestApi.py``
+
+
+[0.5.3] 2018-03-04
 ------------------
 - add documentation for data types in ``neo-python``
 - add intructions on ``build``, ``build .. test``, ``import contract``, and ``testinvoke`` to docs
 - ``BuildNRun`` results now converted to ``ContractParameter`` before printed
 - ``contract {hash}`` no longer throws errors when it is not an ``NEP5`` contract
 - Added method ``AsParameterType`` to ``ContractParameter`` for casting results
+
 
 [0.5.1] 2018-03-02
 ------------------